+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
 # Blocks
 - Added unique sounds for Maple Leaves.
+- Added unique sounds for the Maple wood set. (WIP)
 - Removed duplicate leaves sounds.
 - Fixed Pink Petals incorrectly being assigned a different block SoundType.
 - Removed duplicate flower sounds.